requirements.txt
django==1.8.5
djangorestframework==3.3.0
requests==2.9.1
dj_database_url==0.4.0
psycopg2==2.6.1
pandas==0.17.1
numpy
django-cors-headers
django-filter
gunicorn